requirements.txt
# CEIDG Scraper - Wymagane biblioteki
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.8.0
pandas>=1.5.0
sqlite3  # wbudowane w Python
schedule>=1.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0  # jeśli PostgreSQL
webdriver-manager>=3.8.0  # dla Selenium

# Opcjonalne - dla zaawansowanych funkcji
scrapy>=2.8.0
sqlalchemy>=2.0.0
requests-oauthlib>=1.3.0
cryptography>=40.0.0